refactor and move query to project module
