+++ packages/server/src/project/project.module.ts
@@ -5,8 +5,10 @@ import { PrismaService } from '../prisma/prisma.service';
 import { ProjectPipe } from './project.pipe';
 import { ProjectResolver } from './project.resolver';
 import { UserService } from '../user/user.service';
+import { JwtModule } from '../jwt/jwt.module';
 
 @Module({
+  imports: [JwtModule],
   providers: [ProjectService, PrismaService, ProjectPipe, ProjectResolver, UserService],
   controllers: [ProjectController],
   exports: [ProjectService]
